Modified Logins/Register- React Components/Routes
