+++ api-server-flask/api/routes.py
@@ -126,17 +126,12 @@ class GoogleRegister(Resource):
         'grant_type': 'authorization_code'
         }
         call_for_token = requests.post('https://oauth2.googleapis.com/token', data=payload)
-        print("Call_for_token",call_for_token)
         if call_for_token.status_code == 200:
             response = fetch_user(call_for_token)
-            print("Response", response["status"],response["token"],response["user"])
             if response["status"] == True and response["token"]:
-                print("OKE", response["token"])
                 return {"status":True,"token":response["token"],"user":response["user"]},200
         return {"status":False,"token":None,"user":None},400
             
-               
-                
 
 @rest_api.route('/api/users/register')
 class Register(Resource):
@@ -248,57 +243,9 @@ class LogoutUser(Resource):
 
         return {"success": True}, 200
 
-
-@rest_api.route('/api/sessions/oauth/github/')
-class GitHubLogin(Resource):
-    def get(self):
-        code = request.args.get('code')
-        client_id = BaseConfig.GITHUB_CLIENT_ID
-        client_secret = BaseConfig.GITHUB_CLIENT_SECRET
-        root_url = 'https://github.com/login/oauth/access_token'
-
-        params = { 'client_id': client_id, 'client_secret': client_secret, 'code': code }
-
-        data = requests.post(root_url, params=params, headers={
-            'Content-Type': 'application/x-www-form-urlencoded',
-        })
-
-        response = data._content.decode('utf-8')
-        access_token = response.split('&')[0].split('=')[1]
-
-        user_data = requests.get('https://api.github.com/user', headers={
-            "Authorization": "Bearer " + access_token
-        }).json()
-        
-        user_exists = Users.get_by_username(user_data['login'])
-        if user_exists:
-            user = user_exists
-        else:
-            try:
-                user = Users(username=user_data['login'], email=user_data['email'])
-                user.save()
-            except:
-                user = Users(username=user_data['login'])
-                user.save()
-        
-        user_json = user.toJSON()
-
-        token = jwt.encode({"username": user_json['username'], 'exp': datetime.utcnow() + timedelta(minutes=30)}, BaseConfig.SECRET_KEY)
-        user.set_jwt_auth_active(True)
-        user.save()
-
-        return {"success": True,
-                "user": {
-                    "_id": user_json['_id'],
-                    "email": user_json['email'],
-                    "username": user_json['username'],
-                    "token": token,
-                }}, 200
-
-
 @rest_api.route("/api/users/add_patient",methods=["POST"])
 class UserPatient(Resource):
-    # @token_required
+    @token_required
     def post(self):
         req_data = request.get_json()
         name = req_data["name"]
@@ -338,8 +285,7 @@ class UserPatient(Resource):
 
 @rest_api.route("/api/users/update_patient/<int:patient_id>")
 class UpdatePatient(Resource):
-    
-    # @token_required
+    @token_required
     def put(self, patient_id):
         # Fetch the patient by id from the database
         patient = Patient.query.get(patient_id)
@@ -409,6 +355,7 @@ class UpdatePatient(Resource):
 
 @rest_api.route("/api/users/get_patient/<int:patient_id>")
 class PatientInformation(Resource):
+    @token_required
     def get(self, patient_id):
         # Query the database to get the patient by ID
         patient = Patient.query.get(patient_id)
@@ -441,6 +388,7 @@ class PatientInformation(Resource):
 
 @rest_api.route("/api/users/dashboard")
 class Dashboard(Resource):
+    @token_required
     def get(self):
         # Connect to the SQLite database
         db_path = "/Users/ntuan_195/react-flask-authentication/api-server-flask/api/db.sqlite3"
@@ -501,6 +449,7 @@ class Dashboard(Resource):
 
 @rest_api.route("/api/users/search")
 class GeneralSearch(Resource):
+    @token_required
     def get(self):
         # Get search query parameter
         query = request.args.get('query', default="", type=str)
