+++ react-ui/src/routes/MainRoutes.js
@@ -10,14 +10,9 @@ import AuthGuard from './../utils/route-guard/AuthGuard';
 const DashboardDefault = Loadable(lazy(() => import('../views/dashboard/Default')));
 
 // utilities routing
-const UtilsTypography = Loadable(lazy(() => import('../views/utilities/Typography')));
-const UtilsColor = Loadable(lazy(() => import('../views/utilities/Color')));
-const UtilsShadow = Loadable(lazy(() => import('../views/utilities/Shadow')));
-const UtilsMaterialIcons = Loadable(lazy(() => import('../views/utilities/MaterialIcons')));
-const UtilsTablerIcons = Loadable(lazy(() => import('../views/utilities/TablerIcons')));
+
 const PatientDataComponent = Loadable(lazy(() => import('../views/patient_dashboard/datacomponent')))
 // sample page routing
-const SamplePage = Loadable(lazy(() => import('../views/sample-page')));
 
 //-----------------------|| MAIN ROUTING ||-----------------------//
 
@@ -29,14 +24,6 @@ const MainRoutes = () => {
             path={[
                 '/dashboard/default',
                 '/dashboard/patients',
-
-                '/utils/util-typography',
-                '/utils/util-color',
-                '/utils/util-shadow',
-                '/icons/tabler-icons',
-                '/icons/material-icons',
-
-                '/sample-page'
             ]}
         >
             <MainLayout>
@@ -44,14 +31,6 @@ const MainRoutes = () => {
                     <AuthGuard>
                         <Route path="/dashboard/default" component={DashboardDefault} />
                         <Route path="/dashboard/patients" component={PatientDataComponent} />
-
-                        <Route path="/utils/util-typography" component={UtilsTypography} />
-                        <Route path="/utils/util-color" component={UtilsColor} />
-                        <Route path="/utils/util-shadow" component={UtilsShadow} />
-                        <Route path="/icons/tabler-icons" component={UtilsTablerIcons} />
-                        <Route path="/icons/material-icons" component={UtilsMaterialIcons} />
-
-                        <Route path="/sample-page" component={SamplePage} />
                     </AuthGuard>
                 </Switch>
             </MainLayout>
