+++ react-ui/src/routes/index.js
@@ -4,7 +4,6 @@ import { Redirect, Switch } from 'react-router-dom';
 // routes
 import MainRoutes from './MainRoutes';
 import LoginRoutes from './LoginRoutes';
-import AuthenticationRoutes from './AuthenticationRoutes';
 
 // project imports
 import config from './../config';
@@ -17,7 +16,6 @@ const Routes = () => {
             <Redirect exact from="/" to={config.defaultPath} />
             <React.Fragment>
                 {/* Routes for authentication pages */}
-                <AuthenticationRoutes />
 
                 {/* Route for login */}
                 <LoginRoutes />
