--- react-ui/src/views/sample-page/index.js
@@ -1,24 +1,0 @@
-import React from 'react';
-
-// material-ui
-import { Typography } from '@material-ui/core';
-
-// project imports
-import MainCard from '../../ui-component/cards/MainCard';
-
-//==============================|| SAMPLE PAGE ||==============================//
-
-const SamplePage = () => {
-    return (
-        <MainCard title="Sample Card">
-            <Typography variant="body2">
-                Lorem ipsum dolor sit amen, consenter nipissing eli, sed do elusion tempos incident ut laborers et doolie magna alissa. Ut
-                enif ad minim venice, quin nostrum exercitation illampu laborings nisi ut liquid ex ea commons construal. Duos aube grue
-                dolor in reprehended in voltage veil esse colum doolie eu fujian bulla parian. Exceptive sin ocean cuspidate non president,
-                sunk in culpa qui officiate descent molls anim id est labours.
-            </Typography>
-        </MainCard>
-    );
-};
-
-export default SamplePage;
